Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,11 +6,11 @@
 int main(void)
 {
 	int count = 50;
-	long long fib1 = 1, fib2 = 2;
+	long  fib1 = 1, fib2 = 2;
 
 	for (int i = 0; i < count - 1; i++)
 	{	printf("%lld, ", fib1);
-		long long next = fib1 + fib2;
+		long next = fib1 + fib2;
 
 		fib1 = fib2;
 		fib2 = next;
